moving dataset save fn to shared module
